Split dependent key management into separate module
